% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_feature_weights.R
\name{glmnet_feature_weights}
\alias{glmnet_feature_weights}
\title{Generate feature weights from glmnet models}
\usage{
glmnet_feature_weights(
  y,
  X,
  alpha = 1,
  lambda = 0.01,
  n_rep = 30,
  train_prop = 0.7
)
}
\arguments{
\item{y}{numeric response vector}

\item{X}{matrix or data frame with numeric feature values in columns, instances in rows}

\item{alpha}{glmnet model alpha. \link{0,1} 1 = LASSO (default), 0 = ridge regression, >0 & <1 = elastic net}

\item{lambda}{glmnet model lambda. default = 0.01.}

\item{n_rep}{Num of repetitions. Integer, default = 30}

\item{train_prop}{Proportion of training set to subsample for each repetition. \link{0,1}, default = 0.7}
}
\value{
Returns an object with three items: 1) mean feature weight across all repetitions, 2) data frame with the feature weight from each repetition, and 3) indices of instances selected for each repetition.
}
\description{
This function generates feature weights from glmnet models. Weights are produced from multiple iterations with subsets of training data.
}
\examples{
glmnet_feature_weights(y = y, X = X, alpha = 1, lambda = 0.01, n_rep = 30, train_prop = 0.7)
}
