% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_search_lambda.R
\name{iterate_search_lambda}
\alias{iterate_search_lambda}
\title{Select an optimal lambda value under multiple cross-validation rounds}
\usage{
iterate_search_lambda(y_glm, X_glm, glmnet_parameters_obj)
}
\arguments{
\item{y_glm}{numeric vector with response values}

\item{X_glm}{matrix with features in columns}

\item{glmnet_parameters_obj}{glmnet paramteres object, see "param.LASSO" object loaded with lloydML}
}
\value{
lamda selected most frequently under cross-validation
}
\description{
This function performs cross-validation runs to select an optimal lambda value. Performance metric is minimum mean squared error. Multiple CV runs are performed and the final lambda value is selected based on mode.
}
\examples{
lambda <- iterate_search_lambda(y, X, param)
}
