% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_y.consolidate_performances.R
\name{multi_y.consolidate_performances}
\alias{multi_y.consolidate_performances}
\title{Consolidate multiple performance values}
\usage{
multi_y.consolidate_performances(perfObj, metric_i, P_i)
}
\arguments{
\item{perfObj}{object with performances from multiple response vectors ($performance_objects from multi_y.prediction_and_performance_with_validation_set.wrapper())}

\item{metric_i}{index with target metric: 1 = Pearson's r, 3 = Spearman's rho, 5 = median error, 6 = MSE, 7 = RSS}

\item{P_i}{index with target P-value: 2 = Pearson's p-value, 4 = Spearman's p-value}
}
\value{
Returns a 4-column data frame: col1 = metric, col2 = p-value, col3 = adjusted p-value (FDR method), col4 = p-value factor. Data frame rows contain metrics for multiple response values.
}
\description{
This function will pull the performance metric and P-value of choice and consolidate these values into a single data-frame, with metrics across columns and different response vectors across rows.
}
\examples{
TBD
}
