% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_y.prediction_and_performance.wrapper.R
\name{multi_y.prediction_and_performance_with_validation_set.wrapper}
\alias{multi_y.prediction_and_performance_with_validation_set.wrapper}
\title{Train, apply, and assess performance of prediction models developed with multiple response vectors and a hold-out validation set}
\usage{
multi_y.prediction_and_performance_with_validation_set.wrapper(
  model_type,
  Y,
  X,
  Y.val,
  X.val,
  parameters_obj,
  n_rep = 10,
  prop_per_rep = 0.7,
  parallel = FALSE
)
}
\arguments{
\item{model_type}{algorithm to train model with: "glmnet" or "plsr"}

\item{Y}{matrix with column-wise response values, training set}

\item{X}{matrix with column-wise feature values, training set}

\item{Y.val}{matrix with column-wise response values, validation set}

\item{X.val}{matrix with column-wise feature values, validation set}

\item{parameters_obj}{object with algorithm-specific parameters}

\item{n_rep}{model-building repetitions (i.e. randomly-subsetting training set): integer, default = 10}

\item{prop_per_rep}{proportion of training set to randomly-select for model training each repetition: \link{0, 1}, default = 0.70}

\item{parallel}{Boolean indicating whether to perfom the model repetitions in parallel: default = FALSE}
}
\value{
Returns a set of prediction objects (see train_and_apply_with_validation_set()) and a set of performance objects (see performance.regression()), one for each response vector.
}
\description{
This function takes a matrix of response values and feature values and trains a model for each response columns based on the given feature set. Each model is then applied to a hold-out validation set.
}
\examples{
TBD

}
