% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_glmnet_weights.R
\name{sort_glmnet_weights}
\alias{sort_glmnet_weights}
\title{Sort feature weights from glmnet models}
\usage{
sort_glmnet_weights(glmnet_coef)
}
\arguments{
\item{glmnet_coef}{Glmnet model weights object.}
}
\value{
Returns a 3-column data frame: 1) rank based on absolute feature weight, 2) absolute weight, 3) weight sign (positive or negative).
}
\description{
This function sorts feature weights from glmnet models. Input weight object is generated by: predict(glmnet_model, type = 'coefficients', s = lambda).
}
\examples{
sort_glmnet_weights(glmnet_coef = glmnet_coef)
}
