% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_and_apply_glmnet_model.R
\name{train_and_apply_glmnet_model}
\alias{train_and_apply_glmnet_model}
\title{Train and apply a regularized regression model}
\usage{
train_and_apply_glmnet_model(y_glm, X_glm, newx_glm, glmnet_parameters_obj)
}
\arguments{
\item{y_glm}{response vector - training}

\item{X_glm}{feature matrix - training}

\item{newx_glm}{feature matrix - testing}

\item{glmnet_parameters_obj}{object with glmnet-specific parameters}
}
\value{
Returns an object with three items: 1) vector with predicted response values for the testing instances ($predicted), 2) lambda value selected under CV in the training set ($report_items), and 3) the glmnet model object ($model)
}
\description{
This function will train a regularized regression model and apply it to new data using the 'glmnet' package. This function requires an object with specific parameters. Example parameter objects are attached to this package: param.LASSO, param.ElNet, param.ridge. If requested, lambda will be tuned under CV within the training set. Lambda values to test, CV folds, and CV repetitions are provided in the parameters object.
}
