% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_and_apply_model.R
\name{train_and_apply_model}
\alias{train_and_apply_model}
\title{Train and apply multiple machine learning algorithms}
\usage{
train_and_apply_model(model_type, y, X, newy, newx, parameters_obj)
}
\arguments{
\item{model_type}{algorithm to use: "glment", "plsr"}

\item{y}{response vector - training}

\item{X}{feature matrix - training}

\item{newy}{response vector - testing (plsr only)}

\item{newx}{feature matrix - testing}

\item{parameters_obj}{object with algorithm-specific parameters}
}
\value{
Returns an object with three items: 1) vector with predicted response values for the testing instances ($predicted), 2) reporter items, such as parameters selected ($report_items), and 3) model object ($model)
}
\description{
This function will train a prediction model and apply it to new data.
}
