% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_and_apply_under_CV.R,
%   R/train_and_apply_under_CV.bak.R
\name{train_and_apply_under_CV}
\alias{train_and_apply_under_CV}
\title{Train and apply machine learning algorithms under cross-validation}
\usage{
train_and_apply_under_CV(model_type, y, X, n_fold, n_rep, parameters_obj)

train_and_apply_under_CV(model_type, y, X, n_fold, n_rep, parameters_obj)
}
\arguments{
\item{model_type}{algorithm to use: "glment", "plsr"}

\item{y}{response vector}

\item{X}{feature matrix}

\item{n_fold}{number of CV folds: Integer}

\item{n_rep}{number of repetitions: Integer}

\item{parameters_obj}{object with algorithm-specific parameters}
}
\value{
Returns an object with 2 items: 1) vector with predicted response values ($predicted) and 2) object with reporter items, such as parameters selected, from each CV repetition and CV fold ($report_obj)

Returns an object with 5 items: 1) data frame with predicted value from each CV repetition for each instance ($df_pred), 2) object with reporter items, such as parameters selected, from each CV repetition and CV fold ($report_obj), 3) object with the CV fold assignments for each repetition ($CV_folds), 4) mean predicted value across repetitions for each instance ($pred_mean), and 5) variance in predicted values across reptitions for each instance ($pred_var)
}
\description{
This function will 1) remove instances with NA response values, 2) assign CV folds randomly, 3) iterively subset into training/testing data based on CV folds, and 4) send training/testing data to model training functions. Acts a wrapper around train_and_apply_model().

This function will hold out data by CV fold, train prediction model, and apply it to the held out fold. Acts a wrapper around train_and_apply_model().
}
