% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_CV_folds.R
\name{assign_CV_folds}
\alias{assign_CV_folds}
\title{Assign cross-validation folds}
\usage{
assign_CV_folds(n_instances, n_fold)
}
\arguments{
\item{n_instances}{of instances to assign CV folds: Integer}

\item{n_fold}{of CV folds: Integer}
}
\value{
Returns a vector with fold assignments
}
\description{
This function randomly assigns fold assignments.
}
\examples{
assign_CV_folds(n_instances = 10, n_fold = 3)
> [1] 1 1 1 3 2 1 3 2 2 3
}
