% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_and_apply_with_validation_set.R
\name{train_and_apply_with_validation_set}
\alias{train_and_apply_with_validation_set}
\title{Train and apply machine learning algorithms under cross-validation}
\usage{
train_and_apply_with_validation_set(
  model_type,
  y,
  X,
  X.val,
  n_rep,
  prop_per_rep,
  parameters_obj
)
}
\arguments{
\item{model_type}{algorithm to use: "glment", "plsr"}

\item{y}{response vector}

\item{X}{feature matrix}

\item{n_rep}{number of repetitions: Integer}

\item{parameters_obj}{object with algorithm-specific parameters}

\item{n_fold}{number of CV folds: Integer}
}
\value{
Returns an object with 5 items: 1) data frame with predicted value from each CV repetition for each instance ($df_pred), 2) object with reporter items, such as parameters selected, from each CV repetition and CV fold ($report_obj), 3) object with the CV fold assignments for each repetition ($CV_folds), 4) mean predicted value across repetitions for each instance ($pred_mean), and 5) variance in predicted values across reptitions for each instance ($pred_var)
}
\description{
This function will hold out data by CV fold, train prediction model, and apply it to the held out fold. Acts a wrapper around train_and_apply_model().
}
